Improving default verbosity
This commit aims to reduce/improve the default ('INFO') verbosity.
Most of the INFO logging was changed to DEBUG.

In additional, created a new argument 'enable_logging_info' in the
send_remote_command_batch and send_remote_command methods to
permit to select which remote commands we want to print to INFO verbority
or DEBUG verbority.

Also, use of lazy formatting for logging module.