Acting on SIGTERM
Implement a signal handler for SIGTERM, which will gracefully terminate
the avhclient and allows the context handler's exit branches to be
called. (Including the try/catch/finally's `finally` branch.)
SIGTERMs are usually issued by CI systems to indicate job cancellation
for various reasons, which now the avhclient can act on.